+++ Dfe.PrepareConversions/Dfe.PrepareConversions.Data/Features/PathFor.cs
@@ -19,6 +19,7 @@ public class PathFor
    public static string GetFilterParameters => "/legacy/projects/status";
    public static string AddProjectNote => "/legacy/project/{0}/notes";
    public static string AddSchoolImprovementPlan => "/conversion-project/{0}/school-improvement-plans";
+   public static string GetSchoolImprovementPlans => "/conversion-project/{0}/school-improvement-plans";
    public static string SetExternalApplicationForm => "/conversion-project/{0}/setExternalApplicationForm";
    public static string SetPerformanceData => "/conversion-project/{0}/SetPerformanceData";
    public static string SetIncomingTrust => "/conversion-project/{0}/SetIncomingTrust";
