+++ Dfe.PrepareConversions/Dfe.PrepareConversions/Models/Links.cs
@@ -43,16 +43,16 @@ public static class Links
       public static readonly LinkItem Edit = AddLinkItem(page: "/AnnexB/Edit");
    }
 
-   public static class ImprovementPlans
+   public static class SchoolImprovementPlans
    {
-      public static readonly LinkItem Index = AddLinkItem(backText: "Back", page: "/ImprovementPlans/Index");
-      public static readonly LinkItem WhoArrangedThePlan = AddLinkItem(backText: "Back", page: "/ImprovementPlans/WhoArrangedThePlan");
-      public static readonly LinkItem WhoProvidedThePlan = AddLinkItem(backText: "Back", page: "/ImprovementPlans/WhoProvidedThePlan");
-      public static readonly LinkItem StartDateOfThePlan = AddLinkItem(backText: "Back", page: "/ImprovementPlans/StartDateOfThePlan");
-      public static readonly LinkItem EndDateOfThePlan = AddLinkItem(backText: "Back", page: "/ImprovementPlans/EndDateOfThePlan");
-      public static readonly LinkItem CommentsOnThePlan = AddLinkItem(backText: "Back", page: "/ImprovementPlans/CommentsOnThePlan");
-      public static readonly LinkItem ConfidenceLevelOfThePlan = AddLinkItem(backText: "Back", page: "/ImprovementPlans/ConfidenceLevelOfThePlan");
-      public static readonly LinkItem Summary = AddLinkItem(backText: "Back", page: "/ImprovementPlans/Summary");
+      public static readonly LinkItem Index = AddLinkItem(backText: "Back", page: "/SchoolImprovementPlans/Index");
+      public static readonly LinkItem WhoArrangedThePlan = AddLinkItem(backText: "Back", page: "/SchoolImprovementPlans/WhoArrangedThePlan");
+      public static readonly LinkItem WhoProvidedThePlan = AddLinkItem(backText: "Back", page: "/SchoolImprovementPlans/WhoProvidedThePlan");
+      public static readonly LinkItem StartDateOfThePlan = AddLinkItem(backText: "Back", page: "/SchoolImprovementPlans/StartDateOfThePlan");
+      public static readonly LinkItem EndDateOfThePlan = AddLinkItem(backText: "Back", page: "/SchoolImprovementPlans/EndDateOfThePlan");
+      public static readonly LinkItem CommentsOnThePlan = AddLinkItem(backText: "Back", page: "/SchoolImprovementPlans/CommentsOnThePlan");
+      public static readonly LinkItem ConfidenceLevelOfThePlan = AddLinkItem(backText: "Back", page: "/SchoolImprovementPlans/ConfidenceLevelOfThePlan");
+      public static readonly LinkItem Summary = AddLinkItem(backText: "Back", page: "/SchoolImprovementPlans/Summary");
    }
 
    public static class ExternalApplicationForm
