+++ Dfe.PrepareConversions/Dfe.PrepareConversions/Pages/SchoolImprovementPlans/ConfidenceLevelOfThePlan.cshtml
@@ -1,4 +1,4 @@
-﻿@page "/improvement-plans/{id:int}/confidence-level-of-the-plan"
+﻿@page "/school-improvement-plans/{id:int}/confidence-level-of-the-plan"
 @using Dfe.PrepareConversions.Data.Models.SchoolImprovementPlans
 @using Dfe.Academisation.ExtensionMethods
 @model Dfe.PrepareConversions.Pages.ImprovementPlans.ConfidenceLevelOfThePlanModel
