+++ Dfe.PrepareConversions/Dfe.PrepareConversions/Pages/SchoolImprovementPlans/EndDateOfThePlan.cshtml
@@ -1,4 +1,4 @@
-﻿@page "/improvement-plans/{id:int}/end-date-of-the-plan"
+﻿@page "/school-improvement-plans/{id:int}/end-date-of-the-plan"
 @using Dfe.PrepareConversions.Configuration
 @using Dfe.PrepareConversions.Data.Models.SchoolImprovementPlans
 @using Dfe.Academisation.ExtensionMethods
