+++ Dfe.PrepareConversions/Dfe.PrepareConversions/Pages/SchoolImprovementPlans/Index.cshtml
@@ -1,0 +1,153 @@
+﻿@page "/school-improvement-plans/{id:int}"
+@using Dfe.Academisation.ExtensionMethods
+@using Dfe.PrepareConversions.Configuration
+@using Dfe.PrepareConversions.Data.Models
+@using Dfe.PrepareConversions.Data.Models.SchoolImprovementPlans
+@model Dfe.PrepareConversions.Pages.ImprovementPlans.IndexModel
+@{
+    ViewData["Title"] = "School Improvement Plans";
+}
+
+@section BeforeMain
+{
+    <a asp-page="@Model.ReturnPage" asp-route-id="@Model.ReturnId" class="govuk-back-link">@Links.ProjectList.Index.BackText</a>
+}
+
+@* @if (Model.NewNote)
+{
+   <div class="govuk-notification-banner govuk-notification-banner--success" role="alert" aria-labelledby="govuk-notification-banner-title" data-module="govuk-notification-banner">
+      <div class="govuk-notification-banner__header">
+         <h2 class="govuk-notification-banner__title" id="govuk-notification-banner-title">
+            Success
+         </h2>
+      </div>
+      <div class="govuk-notification-banner__content">
+         <h3 class="govuk-notification-banner__heading" id="project-note-added">
+            Note added
+         </h3>
+      </div>
+   </div>
+} *@
+
+
+<partial name="Shared/_ProjectHeader" model="Model.Project" />
+
+
+<div class="govuk-grid-column-two-thirds-full">
+    <h2 class="govuk-heading-l govuk-!-margin-bottom govuk-!-font-size-36">Interim School Improvement Plans</h2>
+    <a asp-page="@Links.SchoolImprovementPlans.WhoArrangedThePlan.Page" asp-route-id="@Model.Project.Id" role="button" draggable="false" class="govuk-button govuk-button--secondary" data-module="govuk-button">
+        Add School Improvement Plan
+    </a>
+</div>
+
+<div class="govuk-grid-row">
+    <div class="govuk-grid-column-two-thirds">
+        @foreach (var plan in Model.SchoolImprovementPlans)
+        {
+            <div class="govuk-summary-card">
+                <div class="govuk-summary-card__title-wrapper">
+                    <h2 class="govuk-summary-card__title">
+                        Improvement plan created on @plan.CreatedDate.ToString(Constants.DateTimeFormat)
+                    </h2>
+                </div>
+                <div class="govuk-summary-card__content">
+                    <dl class="govuk-summary-list">
+                        <div class="govuk-summary-list__row">
+                            <dt class="govuk-summary-list__key">
+                                Arranged by
+                            </dt>
+                            <dd class="govuk-summary-list__value">
+                                @plan.ArrangedBy.Select(x => x.ToDescription()).Aggregate((partialPhrase, word) =>
+                                         {
+                                             if (SchoolImprovementPlanArranger.Other.ToDescription() == word)
+                                             {
+                                                 return $"{partialPhrase}, {word} - {plan.ArrangedByOther}";
+                                             }
+                                             return $"{partialPhrase}, {word}";
+                                         })
+                            </dd>
+                            <dd class="govuk-summary-list__actions">
+                                <a class="govuk-link" href="#">Change<span class="govuk-visually-hidden"> arranger of the plan</span></a>
+                            </dd>
+                        </div>
+                        <div class="govuk-summary-list__row">
+                            <dt class="govuk-summary-list__key">
+                                Provided by
+                            </dt>
+                            <dd class="govuk-summary-list__value">
+                                @plan.ProvidedBy
+                            </dd>
+                            <dd class="govuk-summary-list__actions">
+                                <a class="govuk-link" href="#">Change<span class="govuk-visually-hidden"> provider of the plan</span></a>
+                            </dd>
+                        </div>
+                        <div class="govuk-summary-list__row">
+                            <dt class="govuk-summary-list__key">
+                                Start date
+                            </dt>
+                            <dd class="govuk-summary-list__value">
+                                @plan.StartDate.Value.ToString(Constants.DateTimeFormat)
+                            </dd>
+                            <dd class="govuk-summary-list__actions">
+                                <a class="govuk-link" href="#">Change<span class="govuk-visually-hidden"> the start date of the plan</span></a>
+                            </dd>
+                        </div>
+                        <div class="govuk-summary-list__row">
+                            <dt class="govuk-summary-list__key">
+                                Expected end date
+                            </dt>
+                            <dd class="govuk-summary-list__value">
+                                @if (plan.ExpectedEndDate.Value == SchoolImprovementPlanExpectedEndDate.Other)
+                                {
+                                    @plan.ExpectedEndDateOther.Value.ToString(Constants.DateTimeFormat)
+                                }
+
+                                @if (plan.ExpectedEndDate.Value == SchoolImprovementPlanExpectedEndDate.ToAdvisoryBoard)
+                                {
+                                    @Model.Project.HeadTeacherBoardDate.Value.ToString(Constants.DateTimeFormat)
+                                }
+
+
+                                @if (plan.ExpectedEndDate.Value == SchoolImprovementPlanExpectedEndDate.ToConversion)
+                                {
+                                    @Model.Project.ProposedAcademyOpeningDate.Value.ToString(Constants.DateTimeFormat)
+                                }
+
+                                @if (plan.ExpectedEndDate.Value == SchoolImprovementPlanExpectedEndDate.Unknown)
+                                {
+                                    @SchoolImprovementPlanExpectedEndDate.Unknown.ToDescription()
+                                }
+                            </dd>
+                            <dd class="govuk-summary-list__actions">
+                                <a class="govuk-link" href="#">Change<span class="govuk-visually-hidden"> the expectedend date of the plan</span></a>
+                            </dd>
+                        </div>
+                        <div class="govuk-summary-list__row">
+                            <dt class="govuk-summary-list__key">
+                                Confidence level
+                            </dt>
+                            <dd class="govuk-summary-list__value">
+                                @plan.ConfidenceLevel.ToDescription()
+                            </dd>
+                            <dd class="govuk-summary-list__actions">
+                                <a class="govuk-link" href="#">Change<span class="govuk-visually-hidden"> the confidence level of the plan</span></a>
+                            </dd>
+                        </div>
+                        <div class="govuk-summary-list__row">
+                            <dt class="govuk-summary-list__key">
+                                Comments
+                            </dt>
+                            <dd class="govuk-summary-list__value">
+                                @plan.PlanComments
+                            </dd>
+                            <dd class="govuk-summary-list__actions">
+                                <a class="govuk-link" href="#">Change<span class="govuk-visually-hidden"> the comments on the plan</span></a>
+                            </dd>
+                        </div>
+                    </dl>
+                </div>
+            </div>
+        }
+
+    </div>
+</div>
