+++ Dfe.PrepareConversions/Dfe.PrepareConversions/Pages/SchoolImprovementPlans/StartDateOfThePlan.cshtml
@@ -1,4 +1,4 @@
-﻿@page "/improvement-plans/{id:int}/start-date-of-the-plan"
+﻿@page "/school-improvement-plans/{id:int}/start-date-of-the-plan"
 @using Dfe.PrepareConversions.Data.Models.SchoolImprovementPlans
 @using Dfe.Academisation.ExtensionMethods
 @model Dfe.PrepareConversions.Pages.ImprovementPlans.StartDateOfThePlanModel
