+++ Dfe.PrepareConversions/Dfe.PrepareConversions/Pages/SchoolImprovementPlans/StartDateOfThePlan.cshtml.cs
@@ -9,11 +9,11 @@ using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
 namespace Dfe.PrepareConversions.Pages.ImprovementPlans;
 
-public class EndDateOfThePlanModel : SchoolImprovementPlanBaseModel, IDateValidationMessageProvider
+public class StartDateOfThePlanModel : SchoolImprovementPlanBaseModel, IDateValidationMessageProvider
 {
    private readonly ErrorService _errorService;
 
-   public EndDateOfThePlanModel(IAcademyConversionProjectRepository repository,
+   public StartDateOfThePlanModel(IAcademyConversionProjectRepository repository,
                            ISession session,
                            ErrorService errorService)
       : base(repository, session)
@@ -21,20 +21,15 @@ public class EndDateOfThePlanModel : SchoolImprovementPlanBaseModel, IDateValida
       _errorService = errorService;
    }
 
-   [BindProperty(Name = "plan-end-date-other", BinderType = typeof(DateInputModelBinder))]
+   [BindProperty(Name = "plan-start-date", BinderType = typeof(DateInputModelBinder))]
    [DateValidation(DateRangeValidationService.DateRange.PastOrFuture)]
-   [Display(Name = "EndDateOther")]
+   [Display(Name = "StartDate")]
    [Required]
-   public DateTime? PlanEndDateOther { get; set; }
-
-   [BindProperty]
-   [Required]
-   [Display(Name = "Expected end date")]
-   public SchoolImprovementPlanExpectedEndDate? ExpectedEndDate{ get; set; }
+   public DateTime? PlanStartDate { get; set; }
 
    public IActionResult OnGet(int id)
    {
-      SetBackLinkModel(Links.ImprovementPlans.StartDateOfThePlan, id);
+      SetBackLinkModel(Links.SchoolImprovementPlans.WhoArrangedThePlan, id);
 
       SchoolImprovementPlan improvementPlan = GetSchoolImprovementPlanFromSession(id);
 
@@ -47,30 +42,22 @@ public class EndDateOfThePlanModel : SchoolImprovementPlanBaseModel, IDateValida
    {
       SchoolImprovementPlan improvementPlan = GetSchoolImprovementPlanFromSession(id);
 
-      if (PlanEndDateOther.HasValue)
+      if (PlanStartDate.HasValue)
       {
-         improvementPlan.ExpectedEndDateOther = PlanEndDateOther.Value;
+         improvementPlan.StartDate = PlanStartDate.Value;
       }
-      improvementPlan.ExpectedEndDate = ExpectedEndDate;
 
       SetSchoolImprovementPlanInSession(id, improvementPlan);
-
-      // Override the validation on the date helper as it is only required when selecting other
-      if (ExpectedEndDate is not SchoolImprovementPlanExpectedEndDate.Other)
-      {
-         this.ViewData.ModelState.Remove("plan-end-date-other");
-      }
 
       _errorService.AddErrors(ModelState.Keys, ModelState);
       if (_errorService.HasErrors()) return OnGet(id);
 
-      return RedirectToPage(Links.ImprovementPlans.ConfidenceLevelOfThePlan.Page, LinkParameters);
+      return RedirectToPage(Links.SchoolImprovementPlans.EndDateOfThePlan.Page, LinkParameters);
    }
 
    private void SetModel(SchoolImprovementPlan schoolImprovementPlan)
    {
-      ExpectedEndDate = schoolImprovementPlan.ExpectedEndDate;
-      PlanEndDateOther = schoolImprovementPlan.ExpectedEndDateOther;
+      PlanStartDate = schoolImprovementPlan.StartDate;
    }
 
    string IDateValidationMessageProvider.SomeMissing(string displayName, IEnumerable<string> missingParts)
@@ -80,6 +67,6 @@ public class EndDateOfThePlanModel : SchoolImprovementPlanBaseModel, IDateValida
 
    string IDateValidationMessageProvider.AllMissing(string displayName)
    {
-      return $"Enter the expected end date of the plan";
+      return $"Enter the start date of the plan";
    }
 }
