+++ Dfe.PrepareConversions/Dfe.PrepareConversions/Pages/SchoolImprovementPlans/WhoArrangedThePlan.cshtml
@@ -1,4 +1,4 @@
-﻿@page "/improvement-plans/{id:int}/who-arranged-the-plan"
+﻿@page "/school-improvement-plans/{id:int}/who-arranged-the-plan"
 @using Dfe.PrepareConversions.Data.Models.SchoolImprovementPlans
 @using Dfe.Academisation.ExtensionMethods
 @model Dfe.PrepareConversions.Pages.ImprovementPlans.WhoArrangedThePlanModel
