+++ Dfe.PrepareConversions/Dfe.PrepareConversions/Pages/SchoolImprovementPlans/_BackLink.cshtml
@@ -4,7 +4,7 @@
    bool.TryParse(Context.Request.Query["obl"], out overideBackLink);
    if (overideBackLink)
    {
-      Model.LinkPage = Links.ImprovementPlans.Summary.Page;
+      Model.LinkPage = Links.SchoolImprovementPlans.Summary.Page;
    }
 }
