+++ Dfe.PrepareConversions/Dfe.PrepareConversions/Pages/Shared/_SubMenu.cshtml
@@ -68,7 +68,7 @@
         </li>
         <li if="@Model.IsSponsored" class="moj-sub-navigation__item">
             <sub-menu-link class="moj-sub-navigation__link"
-                           asp-page="@Links.ImprovementPlans.Index.Page"
+                           asp-page="@Links.SchoolImprovementPlans.Index.Page"
                            asp-route-id="@id">
                 School improvement plans
             </sub-menu-link>
